Refactor admin class
